Add slide right transition
Specifically added two new functions in TPLVisualizer.js: setColorSlide() and clearColorSlide()

The function setColorSlide() changes the background colour of body, and then triggers the slide right animation for 'slide' by adding 'horizTranslate' as to the class list. This creates the illusion of a new search coming in and pushing the old one out.

The function clearColorSlide() simply sets the 'slide' colour to the new background colour, and removes the horizTranslate class allowing the cycle to continue.